Moved HoverPopupWindow to GuiHelpers class, fixed search variable, clean up EditEquip
